requirements.txt
absl_py==0.13.0
albumentations==1.0.3
numpy==1.13.3
opencv_python_headless==4.5.3.56
Pillow==9.0.1
sewar==0.4.4
torch==1.9.0
torchvision==0.10.0
tqdm==4.43.0
